Editing Modules 8-13
I did some quick editing to the last 5 modules. It was mostly fixing formatting issues and a few incorrect website and image references. I also took out the mermaid flow charts from modules 1 and 2 because they were breaking the next button. There is likely a more elegant solution to the problem that should be implemented in the future.